break dev mode hover card to individual components
